Consolidate expected exception checks
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=302490106
