+++ core/src/main/java/com/google/errorprone/bugpatterns/TestExceptionChecker.java
@@ -16,23 +16,32 @@
 
 package com.google.errorprone.bugpatterns;
 
-import static com.google.common.collect.ImmutableList.toImmutableList;
 import static com.google.common.collect.Iterables.getLast;
+import static com.google.common.collect.Iterables.getOnlyElement;
 import static com.google.errorprone.BugPattern.SeverityLevel.WARNING;
 import static com.google.errorprone.matchers.Description.NO_MATCH;
 
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.Lists;
 import com.google.errorprone.BugPattern;
-import com.google.errorprone.BugPattern.ProvidesFix;
-import com.google.errorprone.BugPattern.StandardTags;
 import com.google.errorprone.VisitorState;
+import com.google.errorprone.bugpatterns.BugChecker.MethodTreeMatcher;
 import com.google.errorprone.fixes.SuggestedFix;
 import com.google.errorprone.matchers.Description;
 import com.google.errorprone.matchers.JUnitMatchers;
+import com.google.errorprone.util.ASTHelpers;
+import com.sun.source.tree.ExpressionStatementTree;
+import com.sun.source.tree.ExpressionTree;
 import com.sun.source.tree.MethodTree;
 import com.sun.source.tree.StatementTree;
+import com.sun.source.tree.Tree;
+import com.sun.tools.javac.code.Type;
+import com.sun.tools.javac.tree.JCTree;
+import com.sun.tools.javac.tree.JCTree.JCAnnotation;
+import com.sun.tools.javac.tree.JCTree.JCAssign;
 import com.sun.tools.javac.tree.JCTree.JCExpression;
+import com.sun.tools.javac.tree.JCTree.JCIdent;
+import com.sun.tools.javac.tree.JCTree.JCMethodDecl;
+import com.sun.tools.javac.tree.JCTree.Tag;
+import java.util.Collection;
 import java.util.List;
 
 /** A {@link BugChecker}; see the associated {@link BugPattern} annotation for details. */
@@ -41,40 +50,114 @@ import java.util.List;
     summary =
         "Using @Test(expected=...) is discouraged, since the test will pass if *any* statement in"
             + " the test method throws the expected exception",
-    severity = WARNING,
-    tags = StandardTags.FRAGILE_CODE,
-    providesFix = ProvidesFix.REQUIRES_HUMAN_ATTENTION)
-public class TestExceptionChecker extends AbstractTestExceptionChecker {
+    severity = WARNING)
+public class TestExceptionChecker extends BugChecker implements MethodTreeMatcher {
 
   @Override
-  protected Description handleStatements(
-      MethodTree tree, VisitorState state, JCExpression expectedException, SuggestedFix baseFix) {
-    List<? extends StatementTree> statements = tree.getBody().getStatements();
-    if (statements.size() == 1) {
+  public Description matchMethod(MethodTree tree, VisitorState state) {
+    if (tree.getBody() == null) {
       return NO_MATCH;
     }
-    Description.Builder description = buildDescription(tree);
-    List<SuggestedFix> fixes =
-        Lists.reverse(statements).stream()
-            .filter(t -> !JUnitMatchers.containsTestMethod(t))
-            .map(
-                s ->
-                    buildFix(
-                        state,
-                        SuggestedFix.builder().merge(baseFix),
-                        expectedException,
-                        ImmutableList.of(s)))
-            .collect(toImmutableList());
-    if (!fixes.isEmpty()) {
-      description.addAllFixes(fixes);
-    } else if (!statements.isEmpty()) {
-      description.addFix(
-          buildFix(
-              state,
-              SuggestedFix.builder().merge(baseFix),
-              expectedException,
-              ImmutableList.of(getLast(statements))));
+    SuggestedFix.Builder baseFixBuilder = SuggestedFix.builder();
+    JCExpression expectedException =
+        deleteExpectedException(
+            baseFixBuilder, ((JCMethodDecl) tree).getModifiers().getAnnotations(), state);
+    SuggestedFix baseFix = baseFixBuilder.build();
+    if (expectedException == null) {
+      return NO_MATCH;
     }
-    return description.build();
+    return handleStatements(tree, state, expectedException, baseFix);
+  }
+
+  /**
+   * Handle a method annotated with {@code @Test(expected=...}.
+   *
+   * @param tree the method
+   * @param state the visitor state
+   * @param expectedException the type of expected exception
+   * @param baseFix the base fix
+   */
+  private Description handleStatements(
+      MethodTree tree, VisitorState state, JCExpression expectedException, SuggestedFix baseFix) {
+    return describeMatch(
+        tree,
+        buildFix(
+            state,
+            SuggestedFix.builder().merge(baseFix),
+            expectedException,
+            tree.getBody().getStatements()));
+  }
+
+  private static SuggestedFix buildFix(
+      VisitorState state,
+      SuggestedFix.Builder fix,
+      JCExpression expectedException,
+      Collection<? extends StatementTree> statements) {
+    if (statements.isEmpty()) {
+      return fix.build();
+    }
+    fix.addStaticImport("org.junit.Assert.assertThrows");
+    StringBuilder prefix = new StringBuilder();
+    prefix.append(
+        String.format("assertThrows(%s, () -> ", state.getSourceForNode(expectedException)));
+    if (statements.size() == 1 && getOnlyElement(statements) instanceof ExpressionStatementTree) {
+      ExpressionTree expression =
+          ((ExpressionStatementTree) getOnlyElement(statements)).getExpression();
+      fix.prefixWith(expression, prefix.toString());
+      fix.postfixWith(expression, ")");
+    } else {
+      prefix.append(" {");
+      fix.prefixWith(statements.iterator().next(), prefix.toString());
+      fix.postfixWith(getLast(statements), "});");
+    }
+    return fix.build();
+  }
+
+  /**
+   * Searches the annotation list for {@code @Test(expected=...)}. If found, deletes the exception
+   * attribute from the annotation, and returns its value.
+   */
+  private static JCExpression deleteExpectedException(
+      SuggestedFix.Builder fix, List<JCAnnotation> annotations, VisitorState state) {
+    Type testAnnotation = state.getTypeFromString(JUnitMatchers.JUNIT4_TEST_ANNOTATION);
+    for (JCAnnotation annotationTree : annotations) {
+      if (!ASTHelpers.isSameType(testAnnotation, annotationTree.type, state)) {
+        continue;
+      }
+      com.sun.tools.javac.util.List<JCExpression> arguments = annotationTree.getArguments();
+      for (JCExpression arg : arguments) {
+        if (!arg.hasTag(Tag.ASSIGN)) {
+          continue;
+        }
+        JCAssign assign = (JCAssign) arg;
+        if (assign.lhs.hasTag(Tag.IDENT)
+            && ((JCIdent) assign.lhs).getName().contentEquals("expected")) {
+          if (arguments.size() == 1) {
+            fix.replace(annotationTree, "@Test");
+          } else {
+            removeFromList(fix, state, arguments, assign);
+          }
+          return assign.rhs;
+        }
+      }
+    }
+    return null;
+  }
+
+  /** Deletes an entry and its delimiter from a list. */
+  private static void removeFromList(
+      SuggestedFix.Builder fix, VisitorState state, List<? extends Tree> arguments, Tree tree) {
+    int idx = arguments.indexOf(tree);
+    if (idx == arguments.size() - 1) {
+      fix.replace(
+          state.getEndPosition(arguments.get(arguments.size() - 1)),
+          state.getEndPosition(tree),
+          "");
+    } else {
+      fix.replace(
+          ((JCTree) tree).getStartPosition(),
+          ((JCTree) arguments.get(idx + 1)).getStartPosition(),
+          "");
+    }
   }
 }
