+++ core/src/test/java/com/google/errorprone/bugpatterns/TestExceptionCheckerTest.java
@@ -55,17 +55,19 @@ public class TestExceptionCheckerTest {
             "class ExceptionTest {",
             "  @Test(timeout = 0L)",
             "  public void test() throws Exception {",
-            "    Path p = Paths.get(\"NOSUCH\");",
-            "    Files.readAllBytes(p);",
-            "    assertThrows(IOException.class, () -> Files.readAllBytes(p));",
-            "    assertThat(Files.exists(p)).isFalse();",
+            "    assertThrows(IOException.class, () -> {",
+            "      Path p = Paths.get(\"NOSUCH\");",
+            "      Files.readAllBytes(p);",
+            "      Files.readAllBytes(p);",
+            "      assertThat(Files.exists(p)).isFalse();",
+            "    });",
             "  }",
             "}")
         .doTest();
   }
 
   @Test
-  public void positive_markerException() {
+  public void positive_markerAnnotation() {
     testHelper
         .addInputLines(
             "in/ExceptionTest.java",
@@ -91,16 +93,18 @@ public class TestExceptionCheckerTest {
             "class ExceptionTest {",
             "  @Test",
             "  public void test() throws Exception {",
-            "    Path p = Paths.get(\"NOSUCH\");",
-            "    assertThrows(IOException.class, () -> Files.readAllBytes(p));",
-            "    assertThat(Files.exists(p)).isFalse();",
+            "    assertThrows(IOException.class, () -> {",
+            "      Path p = Paths.get(\"NOSUCH\");",
+            "      Files.readAllBytes(p);",
+            "      assertThat(Files.exists(p)).isFalse();",
+            "    });",
             "  }",
             "}")
         .doTest();
   }
 
   @Test
-  public void negative() {
+  public void oneStatement() {
     testHelper
         .addInputLines(
             "in/ExceptionTest.java",
@@ -113,7 +117,18 @@ public class TestExceptionCheckerTest {
             "    Files.readAllBytes(Paths.get(\"NOSUCH\"));",
             "  }",
             "}")
-        .expectUnchanged()
+        .addOutputLines(
+            "in/ExceptionTest.java",
+            "import static org.junit.Assert.assertThrows;",
+            "import java.io.IOException;",
+            "import java.nio.file.*;",
+            "import org.junit.Test;",
+            "class ExceptionTest {",
+            "  @Test",
+            "  public void test() throws Exception {",
+            "    assertThrows(IOException.class, () -> Files.readAllBytes(Paths.get(\"NOSUCH\")));",
+            "  }",
+            "}")
         .doTest();
   }
 
@@ -129,7 +144,15 @@ public class TestExceptionCheckerTest {
             "  public void test() throws Exception {",
             "  }",
             "}")
-        .expectUnchanged()
+        .addOutputLines(
+            "in/ExceptionTest.java",
+            "import java.io.IOException;",
+            "import org.junit.Test;",
+            "class ExceptionTest {",
+            "  @Test",
+            "  public void test() throws Exception {",
+            "  }",
+            "}")
         .doTest();
   }
 }
