--- core/src/test/java/com/google/errorprone/bugpatterns/TestExceptionRefactoringTest.java
@@ -1,67 +1,0 @@
-/*
- * Copyright 2018 The Error Prone Authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.errorprone.bugpatterns;
-
-import com.google.errorprone.BugCheckerRefactoringTestHelper;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.JUnit4;
-
-/** {@link TestExceptionRefactoring}Test */
-@RunWith(JUnit4.class)
-public class TestExceptionRefactoringTest {
-
-  private final BugCheckerRefactoringTestHelper testHelper =
-      BugCheckerRefactoringTestHelper.newInstance(new TestExceptionRefactoring(), getClass());
-
-  @Test
-  public void positive() {
-    testHelper
-        .addInputLines(
-            "in/ExceptionTest.java",
-            "import static com.google.common.truth.Truth.assertThat;",
-            "import java.io.IOException;",
-            "import java.nio.file.*;",
-            "import org.junit.Test;",
-            "class ExceptionTest {",
-            "  @Test(expected = IOException.class, timeout = 0L)",
-            "  public void test() throws Exception {",
-            "    Path p = Paths.get(\"NOSUCH\");",
-            "    Files.readAllBytes(p);",
-            "    Files.readAllBytes(p);",
-            "  }",
-            "}")
-        .addOutputLines(
-            "out/ExceptionTest.java",
-            "import static com.google.common.truth.Truth.assertThat;",
-            "import static org.junit.Assert.assertThrows;",
-            "import java.io.IOException;",
-            "import java.nio.file.*;",
-            "import org.junit.Test;",
-            "class ExceptionTest {",
-            "  @Test(timeout = 0L)",
-            "  public void test() throws Exception {",
-            "    assertThrows(IOException.class, () -> {",
-            "      Path p = Paths.get(\"NOSUCH\");",
-            "      Files.readAllBytes(p);",
-            "      Files.readAllBytes(p);",
-            "    });",
-            "  }",
-            "}")
-        .doTest();
-  }
-}
